craete new files for the classes
